fix(Google Sheet Node): Refactor Google Sheets (no-changelog) (#15192)
