+++ packages/nodes-base/nodes/Google/Sheet/v2/actions/sheet/read.operation.ts
@@ -1,24 +1,10 @@
-import type {
-	IExecuteFunctions,
-	IDataObject,
-	INodeExecutionData,
-	INodeProperties,
-} from 'n8n-workflow';
+import type { IExecuteFunctions, INodeExecutionData, INodeProperties } from 'n8n-workflow';
 
 import { dataLocationOnSheet, outputFormatting } from './commonDescription';
 import type { GoogleSheet } from '../../helpers/GoogleSheet';
-import type {
-	ILookupValues,
-	RangeDetectionOptions,
-	SheetProperties,
-	SheetRangeData,
-	ValueRenderOption,
-} from '../../helpers/GoogleSheets.types';
-import {
-	getRangeString,
-	prepareSheetData,
-	untilSheetSelected,
-} from '../../helpers/GoogleSheets.utils';
+import type { SheetProperties } from '../../helpers/GoogleSheets.types';
+import { untilSheetSelected } from '../../helpers/GoogleSheets.utils';
+import { readSheet } from '../utils/readOperation';
 
 const combineFiltersOptions: INodeProperties = {
 	displayName: 'Combine Filters',
@@ -41,45 +27,49 @@ const combineFiltersOptions: INodeProperties = {
 	default: 'AND',
 };
 
+export const readFilter: INodeProperties = {
+	displayName: 'Filters',
+	name: 'filtersUI',
+	placeholder: 'Add Filter',
+	type: 'fixedCollection',
+	typeOptions: {
+		multipleValueButtonText: 'Add Filter',
+		multipleValues: true,
+	},
+	default: {},
+	options: [
+		{
+			displayName: 'Filter',
+			name: 'values',
+			values: [
+				{
+					// eslint-disable-next-line n8n-nodes-base/node-param-display-name-wrong-for-dynamic-options
+					displayName: 'Column',
+					name: 'lookupColumn',
+					type: 'options',
+					typeOptions: {
+						loadOptionsDependsOn: ['sheetName.value'],
+						loadOptionsMethod: 'getSheetHeaderRowWithGeneratedColumnNames',
+					},
+					default: '',
+					description:
+						'Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',
+				},
+				{
+					displayName: 'Value',
+					name: 'lookupValue',
+					type: 'string',
+					default: '',
+					hint: 'The column must have this value to be matched',
+				},
+			],
+		},
+	],
+};
+
 export const description: SheetProperties = [
 	{
-		displayName: 'Filters',
-		name: 'filtersUI',
-		placeholder: 'Add Filter',
-		type: 'fixedCollection',
-		typeOptions: {
-			multipleValueButtonText: 'Add Filter',
-			multipleValues: true,
-		},
-		default: {},
-		options: [
-			{
-				displayName: 'Filter',
-				name: 'values',
-				values: [
-					{
-						// eslint-disable-next-line n8n-nodes-base/node-param-display-name-wrong-for-dynamic-options
-						displayName: 'Column',
-						name: 'lookupColumn',
-						type: 'options',
-						typeOptions: {
-							loadOptionsDependsOn: ['sheetName.value'],
-							loadOptionsMethod: 'getSheetHeaderRowWithGeneratedColumnNames',
-						},
-						default: '',
-						description:
-							'Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',
-					},
-					{
-						displayName: 'Value',
-						name: 'lookupValue',
-						type: 'string',
-						default: '',
-						hint: 'The column must have this value to be matched',
-					},
-				],
-			},
-		],
+		...readFilter,
 		displayOptions: {
 			show: {
 				resource: ['sheet'],
@@ -190,97 +180,17 @@ export async function execute(
 		length = items.length;
 	}
 
-	const returnData: INodeExecutionData[] = [];
+	let returnData: INodeExecutionData[] = [];
 
 	for (let itemIndex = 0; itemIndex < length; itemIndex++) {
-		const options = this.getNodeParameter('options', itemIndex, {});
-		const outputFormattingOption =
-			((options.outputFormatting as IDataObject)?.values as IDataObject) || {};
-
-		const dataLocationOnSheetOptions =
-			((options.dataLocationOnSheet as IDataObject)?.values as RangeDetectionOptions) || {};
-
-		if (dataLocationOnSheetOptions.rangeDefinition === undefined) {
-			dataLocationOnSheetOptions.rangeDefinition = 'detectAutomatically';
-		}
-
-		const range = getRangeString(sheetName, dataLocationOnSheetOptions);
-
-		const valueRenderMode = (outputFormattingOption.general ||
-			'UNFORMATTED_VALUE') as ValueRenderOption;
-		const dateTimeRenderOption = (outputFormattingOption.date || 'FORMATTED_STRING') as string;
-
-		const sheetData = (await sheet.getData(
-			range,
-			valueRenderMode,
-			dateTimeRenderOption,
-		)) as SheetRangeData;
-
-		if (sheetData === undefined || sheetData.length === 0) {
-			return [];
-		}
-
-		const {
-			data,
-			headerRow: keyRowIndex,
-			firstDataRow: dataStartRowIndex,
-		} = prepareSheetData(sheetData, dataLocationOnSheetOptions);
-
-		let responseData = [];
-
-		const lookupValues = this.getNodeParameter(
-			'filtersUI.values',
+		returnData = await readSheet.call(
+			this,
+			sheet,
+			sheetName,
 			itemIndex,
-			[],
-		) as ILookupValues[];
-
-		const inputData = data as string[][];
-
-		if (lookupValues.length) {
-			let returnAllMatches;
-			if (nodeVersion < 4.5) {
-				returnAllMatches = options.returnAllMatches === 'returnAllMatches' ? true : false;
-			} else {
-				returnAllMatches = options.returnFirstMatch ? false : true;
-			}
-
-			if (nodeVersion <= 4.1) {
-				for (let i = 1; i < items.length; i++) {
-					const itemLookupValues = this.getNodeParameter(
-						'filtersUI.values',
-						i,
-						[],
-					) as ILookupValues[];
-					if (itemLookupValues.length) {
-						lookupValues.push(...itemLookupValues);
-					}
-				}
-			}
-
-			const combineFilters = this.getNodeParameter('combineFilters', itemIndex, 'OR') as
-				| 'AND'
-				| 'OR';
-
-			responseData = await sheet.lookupValues({
-				inputData,
-				keyRowIndex,
-				dataStartRowIndex,
-				lookupValues,
-				returnAllMatches,
-				combineFilters,
-				nodeVersion,
-			});
-		} else {
-			responseData = sheet.structureArrayDataByColumn(inputData, keyRowIndex, dataStartRowIndex);
-		}
-
-		returnData.push(
-			...responseData.map((item) => {
-				return {
-					json: item,
-					pairedItem: { item: itemIndex },
-				};
-			}),
+			returnData,
+			nodeVersion,
+			items,
 		);
 	}
